Change the way the prior is specified
This creates a dictionary to make it easier for the user to understand
and change the way the prior is chosen. Several scale factor are used
with names corresponding to the parameter and not the distribution as
previously used. This also reorders many of the functions to group
together similar functions.